+++ CHANGELOG.md
@@ -1,7 +1,7 @@
 ## [vNext]
 ### Updated
 - Moved `open` to be before `render`. [Trello 2152](https://trello.com/c/yNzhBkBh)
-- Each running test sends its own rendering task. [Trello 2152](https://trello.com/c/yNzhBkBh)
+- Using server concurrency limitations. [Trello 2152](https://trello.com/c/yNzhBkBh)
 - Separated resource collection from the rendering task. [Trello 2152](https://trello.com/c/yNzhBkBh)
 - Removing resource contents after uploading to the server. [Trello 2152](https://trello.com/c/yNzhBkBh)
 - Calculation viewport size for Android devices. [Trello 2132](https://trello.com/c/CDfbKUV6)